+++ libnavui-androidauto/src/main/java/com/mapbox/androidauto/navigation/audioguidance/CarAudioGuidanceUi.kt
@@ -9,7 +9,6 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.coroutineScope
 import androidx.lifecycle.repeatOnLifecycle
 import com.mapbox.androidauto.R
-import com.mapbox.androidauto.car.FreeDriveActionStrip
 import com.mapbox.androidauto.car.action.MapboxActionProvider
 import com.mapbox.navigation.ui.voice.api.MapboxAudioGuidance
 import kotlinx.coroutines.flow.collect
@@ -17,12 +16,9 @@ import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.launch
 
 /**
- * We're still deciding on a pattern here.
- * Another similar class is [FreeDriveActionStrip].
- *
- * This class creates an action for enabling and disabling audio guidance.
+ * This class creates an action that can control audio guidance.
  */
-class CarAudioGuidanceUi : MapboxActionProvider.ScreenActionProvider {
+class CarAudioGuidanceUi : MapboxActionProvider {
     /**
      * Android auto action for enabling and disabling the car navigation.
      * Attach this to the screen while navigating.